Rename app.py qt_event_loop.py
