+++ napari/util/qt_event_loop.py
